@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:image_scanner_example/page/change_notify_page.dart';
 import 'package:image_scanner_example/page/home_page.dart';
 
 import 'developer/develop_index_page.dart';
@@ -21,13 +22,14 @@ class _IndexPageState extends State<IndexPage> {
           routePage("gallery list", NewHomePage()),
           routePage("save media example", SaveMediaExample()),
           routePage("For Developer page", DeveloperIndexPage()),
+          routePage("Change notify example", ChangeNotifyExample()),
         ],
       ),
     );
   }
 
   Widget routePage(String title, Widget page) {
-    return RaisedButton(
+    return ElevatedButton(
       onPressed: () {
         Navigator.push(
           context,
