@@ -3,10 +3,19 @@ part of '../photo_manager.dart';
 /// asset entity, for entity info.
 class AssetPathEntity {
   /// Obtained by [AssetPathEntity.id], not recommended
-  static Future<AssetPathEntity> fromId(String id,
-      {FilterOptionGroup filterOption}) async {
+  static Future<AssetPathEntity> fromId(
+    String id, {
+    FilterOptionGroup? filterOption,
+    RequestType type = RequestType.common,
+    int albumType = 1,
+  }) async {
+    assert(albumType == 1 || Platform.isIOS || Platform.isMacOS);
     filterOption ??= FilterOptionGroup();
-    final entity = AssetPathEntity()..id = id;
+    final entity = AssetPathEntity()
+      ..id = id
+      ..filterOption = filterOption
+      ..typeInt = type.index
+      ..albumType = 1;
     await entity.refreshPathProperties();
     return entity;
   }
@@ -16,27 +25,32 @@ class AssetPathEntity {
   /// in ios is localIdentifier
   ///
   /// in android is content provider database _id column
-  String id;
+  late final String id;
 
   /// name
   ///
   /// in android is path name
   ///
   /// in ios is photos gallery name
-  String name;
+  late String name;
 
   /// gallery asset count
-  int assetCount;
+  late int assetCount;
 
   /// In iOS: 1: album, 2: folder
   ///
   /// In android: always 1.
-  int albumType;
+  late int albumType;
 
   /// path asset type.
-  RequestType _type;
+  late RequestType _type;
 
-  final FilterOptionGroup filterOption;
+  late FilterOptionGroup filterOption;
+
+  /// The modification time of the latest asset contained in an album.
+  ///
+  /// See [FilterOptionGroup.containsPathModified]
+  DateTime? lastModified;
 
   /// The value used internally by the user.
   /// Used to indicate the value that should be available inside the path.
@@ -62,18 +76,32 @@ class AssetPathEntity {
   /// This path is the path that contains all the assets.
   bool isAll = false;
 
-  AssetPathEntity({this.id, this.name, this.filterOption});
-
   /// Call this method to update the property
-  Future<void> refreshPathProperties({DateTimeCond dateTimeCond}) async {
-    dateTimeCond ??=
-        this.filterOption.dateTimeCond.copyWith(max: DateTime.now());
-    final result = await PhotoManager.fetchPathProperties(this, dateTimeCond);
+  Future<void> refreshPathProperties({
+    bool maxDateTimeToNow = true,
+  }) async {
+    if (maxDateTimeToNow) {
+      filterOption = filterOption.copyWith(
+        createTimeCond: filterOption.createTimeCond.copyWith(
+          max: DateTime.now(),
+        ),
+        updateTimeCond: filterOption.updateTimeCond.copyWith(
+          max: DateTime.now(),
+        ),
+      );
+    }
+
+    final result = await PhotoManager.fetchPathProperties(
+      entity: this,
+      filterOptionGroup: filterOption,
+    );
     if (result != null) {
       this.assetCount = result.assetCount;
       this.name = result.name;
-      this.filterOption.dateTimeCond = dateTimeCond;
       this.isAll = result.isAll;
+      this.type = result.type;
+      this.filterOption = filterOption;
+      this.lastModified = result.lastModified;
     }
   }
 
@@ -93,12 +121,18 @@ class AssetPathEntity {
   }
 
   /// The [start] and [end] like the [String.substring].
-  Future<List<AssetEntity>> getAssetListRange({int start, int end}) async {
+  Future<List<AssetEntity>> getAssetListRange({
+    required int start,
+    required int end,
+  }) async {
     assert(this.albumType == 1, "Just album type can get asset.");
     assert(start >= 0, "The start must better than 0.");
     assert(end > start, "The end must better than start.");
     return PhotoManager._getAssetWithRange(
-        entity: this, start: start, end: end);
+      entity: this,
+      start: start,
+      end: end,
+    );
   }
 
   /// all of asset, It is recommended to use the latest api (pagination) [getAssetListPaged].
@@ -133,19 +167,32 @@ class AssetPathEntity {
 
 /// Used to describe a picture or video
 class AssetEntity {
+  /// see [id]
+  AssetEntity({
+    required this.id,
+    required this.typeInt,
+    required this.width,
+    required this.height,
+    this.duration = 0,
+    this.orientation = 0,
+    this.isFavorite = false,
+    this.title,
+    this.createDtSecond,
+    this.modifiedDateSecond,
+    this.relativePath,
+    double? latitude,
+    double? longitude,
+    this.mimeType,
+  })  : _latitude = latitude,
+        _longitude = longitude;
+
   /// Create from [AssetEntity.id], not recommended.
-  static Future<AssetEntity> fromId(String id) async {
-    final entity = AssetEntity();
-    entity.id = id;
+  static Future<AssetEntity?> fromId(String id) async {
     try {
-      final result = await entity.refreshProperties();
-      if (result == null) {
-        return null;
-      }
+      return await PhotoManager.refreshAssetProperties(id);
     } catch (e) {
       return null;
     }
-    return entity;
   }
 
   /// in android is database _id column
@@ -158,15 +205,12 @@ class AssetEntity {
   /// It is `PHAssetResource.filename` on iOS.
   ///
   /// Nullable in iOS. If you must need it, See [FilterOption.needTitle] or use [titleAsync].
-  String title;
+  String? title;
 
   /// It is [title] in Android.
   ///
   /// It is [PHAsset valueForKey:@"filename"] in iOS.
   Future<String> get titleAsync => _plugin.getTitleAsync(this);
-
-  /// see [id]
-  AssetEntity({this.id});
 
   /// the asset type
   ///
@@ -205,7 +249,7 @@ class AssetEntity {
   /// Gps information when shooting, nullable.
   ///
   /// When the device is android10 or above, always null.
-  double _latitude;
+  double? _latitude;
 
   /// Gps information when shooting, nullable.
   ///
@@ -215,14 +259,14 @@ class AssetEntity {
   /// Gps information when shooting, nullable.
   ///
   /// When the device is android10 or above, always null.
-  set latitude(double latitude) {
+  set latitude(double? latitude) {
     _latitude = latitude;
   }
 
   /// Gps information when shooting, nullable.
   ///
   /// When the device is android10 or above, always null.
-  double _longitude;
+  double? _longitude;
 
   /// Gps information when shooting, nullable.
   ///
@@ -232,9 +276,14 @@ class AssetEntity {
   /// Gps information when shooting, nullable.
   ///
   /// When the device is android10 or above, always null.
-  set longitude(double longitude) {
+  set longitude(double? longitude) {
     _longitude = longitude;
   }
+
+  /// Whether this asset is locally available.
+  ///
+  /// Defaults to true, and it'll always be true on Android.
+  Future<bool> get isLocallyAvailable => PhotoManager._isLocallyAvailable(id);
 
   /// Get latitude and longitude from MediaStore(android) / Photos(iOS).
   ///
@@ -245,13 +294,17 @@ class AssetEntity {
     return _plugin.getLatLngAsync(this);
   }
 
-  /// if you need upload file ,then you can use the file, nullable.
-  Future<File> get file async => PhotoManager._getFileWithId(this.id);
+  /// If you need upload file ,then you can use the file, nullable.
+  ///
+  /// If you need to see the loading status, look at the [loadFile].
+  Future<File?> get file async => PhotoManager._getFileWithId(this.id);
 
   /// This contains all the EXIF information, but in contrast, `Image` widget may not be able to display pictures.
   ///
-  /// Usually, you can use the [file] attribute
-  Future<File> get originFile async =>
+  /// Usually, you can use the [file] attribute.
+  ///
+  /// If you need to see the loading status, look at the [loadFile].
+  Future<File?> get originFile async =>
       PhotoManager._getFileWithId(id, isOrigin: true);
 
   /// The asset's bytes.
@@ -260,30 +313,44 @@ class AssetEntity {
   ///
   /// The property will be remove in 0.5.0.
   @Deprecated("Use originBytes instead")
-  Future<Uint8List> get fullData => PhotoManager._getFullDataWithId(id);
+  Future<Uint8List?> get fullData => PhotoManager._getFullDataWithId(id);
 
   /// The raw data stored in the device, the data may be large.
   ///
   /// This property is not recommended for video types.
-  Future<Uint8List> get originBytes => PhotoManager._getOriginBytes(this);
+  Future<Uint8List?> get originBytes => PhotoManager._getOriginBytes(this);
 
-  /// thumb data , for display
-  Future<Uint8List> get thumbData => thumbDataWithSize(150, 150);
+  /// The thumb data of asset, use it to display.
+  ///
+  /// If you need to see the loading status, look at the [thumbDataWithSize] or [thumbDataWithOption].
+  Future<Uint8List?> get thumbData => thumbDataWithSize(150, 150);
 
   /// get thumb with size
-  Future<Uint8List> thumbDataWithSize(
+  Future<Uint8List?> thumbDataWithSize(
     int width,
     int height, {
     ThumbFormat format = ThumbFormat.jpeg,
     int quality = 100,
-  }) {
+    PMProgressHandler? progressHandler,
+  }) async {
     assert(width > 0 && height > 0, "The width and height must better 0.");
-    assert(format != null, "The format must not be null.");
     assert(quality > 0 && quality <= 100, "The quality must between 0 and 100");
 
     /// Return null if asset is audio or other type, because they don't have such a thing.
     if (type == AssetType.audio || type == AssetType.other) {
       return null;
+    }
+
+    if (Platform.isIOS || Platform.isMacOS) {
+      return thumbDataWithOption(
+        ThumbOption.ios(
+          width: width,
+          height: height,
+          format: format,
+          quality: quality,
+        ),
+        progressHandler: progressHandler,
+      );
     }
 
     return thumbDataWithOption(
@@ -293,13 +360,27 @@ class AssetEntity {
         format: format,
         quality: quality,
       ),
+      progressHandler: progressHandler,
     );
   }
 
-  /// get thumb with size
-  Future<Uint8List> thumbDataWithOption(
-    ThumbOption option,
-  ) {
+  /// Use the method to load file.
+  Future<File?> loadFile({
+    bool isOrigin = true,
+    PMProgressHandler? progressHandler,
+  }) async {
+    return PhotoManager._getFileWithId(
+      id,
+      isOrigin: isOrigin,
+      progressHandler: progressHandler,
+    );
+  }
+
+  /// Get thumb with size of option.
+  Future<Uint8List?> thumbDataWithOption(
+    ThumbOption option, {
+    PMProgressHandler? progressHandler,
+  }) async {
     assert(() {
       option.checkAssert();
       return true;
@@ -313,17 +394,18 @@ class AssetEntity {
     return PhotoManager._getThumbDataWithOption(
       id,
       option,
+      progressHandler,
     );
   }
 
-  /// if not video ,duration is null
-  Duration get videoDuration => Duration(seconds: duration ?? 0);
+  /// if not video, duration is 0
+  Duration get videoDuration => Duration(seconds: duration);
 
   /// nullable, if the manager is null.
   Size get size => Size(width.toDouble(), height.toDouble());
 
-  /// unix timestamp of asset, milliseconds
-  int createDtSecond;
+  /// unix timestamp second of asset
+  int? createDtSecond;
 
   /// create time of asset
   DateTime get createDateTime {
@@ -331,8 +413,8 @@ class AssetEntity {
     return DateTime.fromMillisecondsSinceEpoch(sec * 1000);
   }
 
-  /// second of modified.
-  int modifiedDateSecond;
+  /// unix time second of modified.
+  int? modifiedDateSecond;
 
   DateTime get modifiedDateTime {
     final sec = modifiedDateSecond ?? 0;
@@ -342,12 +424,13 @@ class AssetEntity {
   /// If the asset is deleted, return false.
   Future<bool> get exists => PhotoManager._assetExistsWithId(id);
 
-  /// The url is provided to some video player. Such as [flutter_ijkplayer](https://pub.dev/packages/flutter_ijkplayer)
+  /// The url is provided to some video player.
+  /// Such as [flutter_ijkplayer](https://pub.dev/packages/flutter_ijkplayer)
   ///
   /// It is such as `file:///var/mobile/Media/DCIM/118APPLE/IMG_8371.MOV` in iOS.
   ///
   /// Android: `content://media/external/video/media/894857`
-  Future<String> getMediaUrl() {
+  Future<String?> getMediaUrl() async {
     if (type == AssetType.video || type == AssetType.audio) {
       return PhotoManager._getMediaUrl(this);
     }
@@ -371,11 +454,18 @@ class AssetEntity {
   /// In androidQ or higher: it is `MediaStore.MediaColumns.RELATIVE_PATH`.
   ///
   /// API 28 or lower: it is `MediaStore.MediaColumns.DATA` parent path.
-  String relativePath;
+  String? relativePath;
+
+  /// MimeType of the asset.
+  ///
+  /// In Android, it reads the value in [MediaStore](https://developer.android.com/reference/android/provider/MediaStore.MediaColumns#MIME_TYPE) without guarantee of accuracy.
+  ///
+  /// In iOS, it it always null.
+  String? mimeType;
 
   /// refreshProperties
-  Future<AssetEntity> refreshProperties() async {
-    return PhotoManager.refreshAssetProperties(this);
+  Future<AssetEntity?> refreshProperties() async {
+    return PhotoManager.refreshAssetProperties(this.id);
   }
 
   @override
@@ -392,16 +482,16 @@ class AssetEntity {
   }
 
   @override
-  String toString() {
-    return "AssetEntity{ id:$id , type: $type}";
-  }
+  String toString() => "AssetEntity (id:$id , type: $type)";
 }
 
 /// Longitude and latitude
 class LatLng {
-  /// longitude
-  double longitude;
+  const LatLng({required this.latitude, required this.longitude});
 
   /// latitude
-  double latitude;
+  final double? latitude;
+
+  /// longitude
+  final double? longitude;
 }
