@@ -2,20 +2,17 @@ import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:photo_manager/photo_manager.dart';
-import 'package:photo_manager/src/utils/convert_utils.dart';
 
-/// Any method of this class is not recommended to be called directly outside the library.
+import 'utils/convert_utils.dart';
+
+/// Any method of this class is not recommended to be called directly outside
+/// the library.
 class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
-  static Plugin _plugin;
+  static final Plugin _plugin = Plugin._();
 
-  /// Any method of this class is not recommended to be called directly outside the library.
-  factory Plugin() {
-    _plugin ??= Plugin._();
-    return _plugin;
-  }
+  factory Plugin() => _plugin;
 
   Plugin._();
 
@@ -23,14 +20,14 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
   Future<List<AssetPathEntity>> getAllGalleryList({
     int type = 0,
     bool hasAll = true,
-    FilterOptionGroup optionGroup,
-    bool onlyAll,
+    bool onlyAll = false,
+    required FilterOptionGroup optionGroup,
   }) async {
-    final result = await _channel.invokeMethod("getGalleryList", {
-      "type": type,
-      "hasAll": hasAll,
-      "onlyAll": onlyAll,
-      "option": optionGroup.toMap(),
+    final result = await _channel.invokeMethod('getGalleryList', {
+      'type': type,
+      'hasAll': hasAll,
+      'onlyAll': onlyAll,
+      'option': optionGroup.toMap(),
     });
     if (result == null) {
       return [];
@@ -42,9 +39,13 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
     );
   }
 
-  /// request permission.
-  Future<bool> requestPermission() async {
-    return (await _channel.invokeMethod("requestPermission")) == 1;
+  Future<int> requestPermissionExtend(
+    PermisstionRequestOption requestOption,
+  ) async {
+    return await _channel.invokeMethod(
+      'requestPermissionExtend',
+      requestOption.toMap(),
+    );
   }
 
   /// Use pagination to get album content.
@@ -53,14 +54,14 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
     int page = 0,
     int pageCount = 15,
     int type = 0,
-    FilterOptionGroup optionGroup,
+    required FilterOptionGroup optionGroup,
   }) async {
-    final result = await _channel.invokeMethod("getAssetWithGalleryId", {
-      "id": id,
-      "page": page,
-      "pageCount": pageCount,
-      "type": type,
-      "option": optionGroup.toMap(),
+    final result = await _channel.invokeMethod('getAssetWithGalleryId', {
+      'id': id,
+      'page': page,
+      'pageCount': pageCount,
+      'type': type,
+      'option': optionGroup.toMap(),
     });
 
     return ConvertUtils.convertToAssetList(result);
@@ -69,77 +70,96 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
   /// Asset in the specified range.
   Future<List<AssetEntity>> getAssetWithRange(
     String id, {
-    int typeInt,
-    int start,
-    int end,
-    FilterOptionGroup optionGroup,
+    required int typeInt,
+    required int start,
+    required int end,
+    required FilterOptionGroup optionGroup,
   }) async {
-    final Map map = await _channel.invokeMethod("getAssetListWithRange", {
-      "galleryId": id,
-      "type": typeInt,
-      "start": start,
-      "end": end,
-      "option": optionGroup.toMap(),
+    final Map map = await _channel.invokeMethod('getAssetListWithRange', {
+      'galleryId': id,
+      'type': typeInt,
+      'start': start,
+      'end': end,
+      'option': optionGroup.toMap(),
     });
 
     return ConvertUtils.convertToAssetList(map);
   }
 
+  void _injectParams(Map params, PMProgressHandler? progressHandler) {
+    if (progressHandler != null) {
+      params['progressHandler'] = progressHandler.channelIndex;
+    }
+  }
+
   /// Get thumb of asset id.
-  Future<Uint8List> getThumb({
-    @required String id,
-    @required ThumbOption option,
+  Future<Uint8List?> getThumb({
+    required String id,
+    required ThumbOption option,
+    PMProgressHandler? progressHandler,
   }) {
-    return _channel.invokeMethod("getThumb", {
-      "id": id,
-      "option": option.toMap(),
-    });
+    final params = {
+      'id': id,
+      'option': option.toMap(),
+    };
+    _injectParams(params, progressHandler);
+    return _channel.invokeMethod('getThumb', params);
   }
 
-  Future<Uint8List> getOriginBytes(String id) async {
-    return _channel.invokeMethod("getOriginBytes", {"id": id});
+  Future<Uint8List?> getOriginBytes(
+    String id, {
+    PMProgressHandler? progressHandler,
+  }) {
+    final params = {'id': id};
+    _injectParams(params, progressHandler);
+    return _channel.invokeMethod('getOriginBytes', params);
   }
 
-  Future<void> releaseCache() async {
-    await _channel.invokeMethod("releaseMemCache");
+  Future<void> releaseCache() {
+    return _channel.invokeMethod('releaseMemCache');
   }
 
-  Future<String> getFullFile(String id, {bool isOrigin}) async {
-    return _channel.invokeMethod("getFullFile", {
-      "id": id,
-      "isOrigin": isOrigin,
-    });
+  Future<String?> getFullFile(
+    String id, {
+    required bool isOrigin,
+    PMProgressHandler? progressHandler,
+  }) async {
+    final params = {'id': id, 'isOrigin': isOrigin};
+    _injectParams(params, progressHandler);
+    return _channel.invokeMethod('getFullFile', params);
   }
 
-  Future<void> setLog(bool isLog) async {
-    return _channel.invokeMethod("log", isLog);
+  Future<void> setLog(bool isLog) {
+    return _channel.invokeMethod('log', isLog);
   }
 
-  void openSetting() {
-    _channel.invokeMethod("openSetting");
+  Future<void> openSetting() {
+    return _channel.invokeMethod('openSetting');
   }
 
-  Future<Map> fetchPathProperties(
-      String id, int type, FilterOptionGroup optionGroup) async {
+  /// Nullable
+  Future<Map?> fetchPathProperties(
+    String id,
+    int type,
+    FilterOptionGroup optionGroup,
+  ) {
     return _channel.invokeMethod(
-      "fetchPathProperties",
+      'fetchPathProperties',
       {
-        "id": id,
-        "timestamp": 0,
-        "type": type,
-        "option": optionGroup.toMap(),
+        'id': id,
+        'timestamp': 0,
+        'type': type,
+        'option': optionGroup.toMap(),
       },
     );
   }
 
-  void notifyChange({bool start}) {
-    _channel.invokeMethod("notify", {
-      "notify": start,
-    });
+  void notifyChange({required bool start}) {
+    _channel.invokeMethod('notify', {'notify': start});
   }
 
   Future<void> forceOldApi() async {
-    await _channel.invokeMethod("forceOldApi");
+    await _channel.invokeMethod('forceOldApi');
   }
 
   Future<bool> deleteWithId(String id) async {
@@ -149,129 +169,145 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
 
   Future<List<String>> deleteWithIds(List<String> ids) async {
     final List<dynamic> deleted =
-        (await _channel.invokeMethod("deleteWithIds", {"ids": ids}));
+        (await _channel.invokeMethod('deleteWithIds', {'ids': ids}));
     return deleted.cast<String>();
   }
 
-  Future<AssetEntity> saveImage(Uint8List uint8list,
-      {String title, String desc = ""}) async {
-    title ??= "image_${DateTime.now().millisecondsSinceEpoch / 1000}";
+  Future<AssetEntity?> saveImage(
+    Uint8List data, {
+    String? title,
+    String? desc,
+    String? relativePath,
+  }) async {
+    title ??= 'image_${DateTime.now().millisecondsSinceEpoch / 1000}.jpg';
 
     final result = await _channel.invokeMethod(
-      "saveImage",
+      'saveImage',
       {
-        "image": uint8list,
-        "title": title,
-        "desc": desc,
+        'image': data,
+        'title': title,
+        'desc': desc ?? '',
+        'relativePath': relativePath,
       },
     );
 
     return ConvertUtils.convertToAsset(result);
   }
 
-  Future<AssetEntity> saveImageWithPath(String path,
-      {String title, String desc = ""}) async {
+  Future<AssetEntity?> saveImageWithPath(
+    String path, {
+    String? title,
+    String? desc,
+    String? relativePath,
+  }) async {
     final file = File(path);
     if (!file.existsSync()) {
-      assert(file.existsSync(), "file must exists");
+      assert(file.existsSync(), 'file must exists');
       return null;
     }
 
-    title ??= "image_${DateTime.now().millisecondsSinceEpoch / 1000}.jpg";
+    title ??= 'image_${DateTime.now().millisecondsSinceEpoch / 1000}.jpg';
 
     final result = await _channel.invokeMethod(
-      "saveImageWithPath",
+      'saveImageWithPath',
       {
-        "path": path,
-        "title": title,
-        "desc": desc,
+        'path': path,
+        'title': title,
+        'desc': desc ?? '',
+        'relativePath': relativePath,
       },
     );
 
     return ConvertUtils.convertToAsset(result);
   }
 
-  Future<AssetEntity> saveVideo(
+  Future<AssetEntity?> saveVideo(
     File file, {
-    String title,
-    String desc = "",
+    String? title,
+    String? desc,
+    String? relativePath,
   }) async {
     if (!file.existsSync()) {
-      assert(file.existsSync(), "file must exists");
+      assert(file.existsSync(), 'file must exists');
       return null;
     }
     final result = await _channel.invokeMethod(
-      "saveVideo",
+      'saveVideo',
       {
-        "path": file.absolute.path,
-        "title": title,
-        "desc": desc,
+        'path': file.absolute.path,
+        'title': title,
+        'desc': desc ?? '',
+        'relativePath': relativePath,
       },
     );
     return ConvertUtils.convertToAsset(result);
   }
 
-  Future<bool> assetExistsWithId(String id) {
-    return _channel.invokeMethod("assetExists", {"id": id});
+  Future<bool> assetExistsWithId(String id) async {
+    return await _channel.invokeMethod('assetExists', {'id': id});
   }
 
   Future<String> getSystemVersion() async {
-    return _channel.invokeMethod("systemVersion");
+    return await _channel.invokeMethod('systemVersion');
   }
 
   Future<LatLng> getLatLngAsync(AssetEntity assetEntity) async {
     if (Platform.isAndroid) {
       final version = int.parse(await getSystemVersion());
       if (version >= 29) {
-        final map = await _channel
-            .invokeMethod("getLatLngAndroidQ", {"id": assetEntity.id});
-        if (map is Map) {
-          /// 将返回的数据传入map
-          return LatLng()
-            ..latitude = map["lat"]
-            ..longitude = map["lng"];
-        }
+        final Map? map = await _channel.invokeMethod(
+          'getLatLngAndroidQ',
+          {'id': assetEntity.id},
+        );
+
+        /// 将返回的数据传入map
+        return LatLng(latitude: map?['lat'], longitude: map?['lng']);
       }
     }
-    return LatLng()
-      ..latitude = assetEntity.latitude
-      ..longitude = assetEntity.longitude;
+    return LatLng(
+      latitude: assetEntity.latitude,
+      longitude: assetEntity.longitude,
+    );
   }
 
-  Future<bool> cacheOriginBytes(bool cache) {
-    return _channel.invokeMethod("cacheOriginBytes");
+  Future<bool> cacheOriginBytes(bool cache) async {
+    return (await _channel.invokeMethod("cacheOriginBytes", cache)) == true;
   }
 
   Future<String> getTitleAsync(AssetEntity assetEntity) async {
     assert(Platform.isAndroid || Platform.isIOS || Platform.isMacOS);
     if (Platform.isAndroid) {
-      return assetEntity.title;
+      return assetEntity.title!;
     }
 
     if (Platform.isIOS || Platform.isMacOS) {
-      return _channel.invokeMethod("getTitleAsync", {"id": assetEntity.id});
+      return await _channel.invokeMethod(
+        'getTitleAsync',
+        {'id': assetEntity.id},
+      );
     }
 
-    return "";
+    return '';
   }
 
-  Future<String> getMediaUrl(AssetEntity assetEntity) {
-    return _channel.invokeMethod("getMediaUrl", {
-      "id": assetEntity.id,
-      "type": assetEntity.typeInt,
+  Future<String?> getMediaUrl(AssetEntity assetEntity) {
+    return _channel.invokeMethod('getMediaUrl', {
+      'id': assetEntity.id,
+      'type': assetEntity.typeInt,
     });
   }
 
   Future<List<AssetPathEntity>> getSubPathEntities(
-      AssetPathEntity pathEntity) async {
-    final result = await _channel.invokeMethod("getSubPath", {
-      "id": pathEntity.id,
-      "type": pathEntity.type.value,
-      "albumType": pathEntity.albumType,
-      "option": pathEntity.filterOption.toMap(),
+    AssetPathEntity pathEntity,
+  ) async {
+    final result = await _channel.invokeMethod('getSubPath', {
+      'id': pathEntity.id,
+      'type': pathEntity.type.value,
+      'albumType': pathEntity.albumType,
+      'option': pathEntity.filterOption.toMap(),
     });
 
-    final items = result["list"];
+    final items = result['list'];
 
     return ConvertUtils.convertPath(
       items,
@@ -280,17 +316,21 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
     );
   }
 
-  Future<AssetEntity> copyAssetToGallery(
-      AssetEntity asset, AssetPathEntity pathEntity) async {
+  Future<AssetEntity?> copyAssetToGallery(
+    AssetEntity asset,
+    AssetPathEntity pathEntity,
+  ) async {
     if (pathEntity.isAll) {
-      assert(pathEntity.isAll,
-          "You don't need to copy the asset into the album containing all the pictures.");
+      assert(
+        pathEntity.isAll,
+        'You don\'t need to copy the asset into the album containing all the pictures.',
+      );
       return null;
     }
 
-    final result = await _channel.invokeMethod("copyAsset", {
-      "assetId": asset.id,
-      "galleryId": pathEntity.id,
+    final result = await _channel.invokeMethod('copyAsset', {
+      'assetId': asset.id,
+      'galleryId': pathEntity.id,
     });
 
     if (result == null) {
@@ -301,30 +341,31 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
   }
 
   Future<bool> iosDeleteCollection(AssetPathEntity path) async {
-    final result = await _channel.invokeMethod("deleteAlbum", {
-      "id": path.id,
-      "type": path.albumType,
+    final result = await _channel.invokeMethod('deleteAlbum', {
+      'id': path.id,
+      'type': path.albumType,
     });
-    if (result["errorMsg"] != null) {
-      print(result["errorMsg"]);
+    if (result['errorMsg'] != null) {
+      print(result['errorMsg']);
       return false;
     }
     return true;
   }
 
-  Future<bool> favoriteAsset(String id, bool favorite) {
-    return _channel.invokeMethod("favoriteAsset", {
-      "id": id,
-      "favorite": favorite,
-    });
+  Future<bool> favoriteAsset(String id, bool favorite) async {
+    return (await _channel.invokeMethod(
+          'favoriteAsset',
+          {'id': id, 'favorite': favorite},
+        )) ==
+        true;
   }
 
-  Future<bool> androidRemoveNoExistsAssets() {
-    return _channel.invokeMethod("removeNoExistsAssets");
+  Future<bool> androidRemoveNoExistsAssets() async {
+    return (await _channel.invokeMethod('removeNoExistsAssets')) == true;
   }
 
-  Future getPropertiesFromAssetEntity(String id) async {
-    return _channel.invokeMethod('getPropertiesFromAssetEntity', {"id": id});
+  Future<Map?> getPropertiesFromAssetEntity(String id) {
+    return _channel.invokeMethod('getPropertiesFromAssetEntity', {'id': id});
   }
 
   Future ignorePermissionCheck(bool ignore) async {
@@ -340,14 +381,20 @@ class Plugin with BasePlugin, IosPlugin, AndroidPlugin {
   }
 
   Future<void> requestCacheAssetsThumb(
-      List<String> ids, ThumbOption option) async {
-    assert(ids != null);
+    List<String> ids,
+    ThumbOption option,
+  ) {
     assert(ids.isNotEmpty);
-    assert(option != null);
-    await _channel.invokeMethod('requestCacheAssetsThumb', {
+    return _channel.invokeMethod('requestCacheAssetsThumb', {
       'ids': ids,
       'option': option.toMap(),
     });
+  }
+
+  Future<void> presentLimited() async {
+    if (Platform.isIOS) {
+      await _channel.invokeMethod('presentLimited');
+    }
   }
 }
 
@@ -356,60 +403,73 @@ mixin BasePlugin {
 }
 
 mixin IosPlugin on BasePlugin {
-  Future<AssetPathEntity> iosCreateFolder(
-      String name, bool isRoot, AssetPathEntity parent) async {
+  Future<bool> isLocallyAvailable(String id) async {
+    if (Platform.isAndroid) {
+      return true;
+    }
+    return await _channel.invokeMethod('isLocallyAvailable', {'id': id});
+  }
+
+  Future<AssetPathEntity?> iosCreateFolder(
+    String name,
+    bool isRoot,
+    AssetPathEntity? parent,
+  ) async {
     final map = {
-      "name": name,
-      "isRoot": isRoot,
+      'name': name,
+      'isRoot': isRoot,
     };
     if (!isRoot && parent != null) {
-      map["folderId"] = parent.id;
+      map['folderId'] = parent.id;
     }
     final result = await _channel.invokeMethod(
-      "createFolder",
+      'createFolder',
       map,
     );
     if (result == null) {
       return null;
     }
 
-    if (result["errorMsg"] != null) {
-      print("errorMsg");
+    if (result['errorMsg'] != null) {
+      print('errorMsg');
       return null;
     }
 
     return AssetPathEntity()
-      ..id = result["id"]
+      ..id = result['id']
       ..name = name
       ..isAll = false
       ..assetCount = 0
       ..albumType = 2;
   }
 
-  Future<AssetPathEntity> iosCreateAlbum(
-      String name, bool isRoot, AssetPathEntity parent) async {
+  Future<AssetPathEntity?> iosCreateAlbum(
+    String name,
+    bool isRoot,
+    AssetPathEntity? parent,
+  ) async {
     final map = {
-      "name": name,
-      "isRoot": isRoot,
+      'name': name,
+      'isRoot': isRoot,
     };
     if (!isRoot && parent != null) {
-      map["folderId"] = parent.id;
+      map['folderId'] = parent.id;
     }
     final result = await _channel.invokeMethod(
-      "createAlbum",
+      'createAlbum',
       map,
     );
     if (result == null) {
       return null;
     }
 
-    if (result["errorMsg"] != null) {
-      print("errorMsg");
+    if (result['errorMsg'] != null) {
+      print('errorMsg');
       return null;
     }
 
     return AssetPathEntity()
-      ..id = result["id"]
+      ..id = result['id']
       ..name = name
       ..isAll = false
       ..assetCount = 0
@@ -417,17 +477,19 @@ mixin IosPlugin on BasePlugin {
   }
 
   Future<bool> iosRemoveInAlbum(
-      List<AssetEntity> entities, AssetPathEntity path) async {
+    List<AssetEntity> entities,
+    AssetPathEntity path,
+  ) async {
     final result = await _channel.invokeMethod(
-      "removeInAlbum",
+      'removeInAlbum',
       {
-        "assetId": entities.map((e) => e.id).toList(),
-        "pathId": path.id,
+        'assetId': entities.map((e) => e.id).toList(),
+        'pathId': path.id,
       },
     );
 
-    if (result["msg"] != null) {
-      print("cannot remove, cause by: ${result["msg"]}");
+    if (result['msg'] != null) {
+      print('cannot remove, cause by: ${result['msg']}');
       return false;
     }
 
@@ -437,10 +499,12 @@ mixin IosPlugin on BasePlugin {
 
 mixin AndroidPlugin on BasePlugin {
   Future<bool> androidMoveAssetToPath(
-      AssetEntity entity, AssetPathEntity target) async {
-    final result = await _channel.invokeMethod("moveAssetToPath", {
-      "assetId": entity.id,
-      "albumId": target.id,
+    AssetEntity entity,
+    AssetPathEntity target,
+  ) async {
+    final result = await _channel.invokeMethod('moveAssetToPath', {
+      'assetId': entity.id,
+      'albumId': target.id,
     });
 
     print(result);
