@@ -7,8 +7,7 @@ import java.util.List;
  */
 
 public interface PermissionsListener {
+    void onDenied(List<String> deniedPermissions, List<String> grantedPermissions);
 
-  void onDenied(List<String> deniedPermissions, List<String> grantedPermissions);
-
-  void onGranted();
+    void onGranted();
 }
