@@ -1,15 +1,11 @@
 package top.kikt.imagescannerexample
 
 import android.os.Bundle
+import io.flutter.embedding.android.FlutterActivity
 
-import io.flutter.app.FlutterActivity
-import io.flutter.plugins.GeneratedPluginRegistrant
-
-class MainActivity() : FlutterActivity() {
+class MainActivity : FlutterActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        GeneratedPluginRegistrant.registerWith(this)
-
         CrashHandler.initHandler(this)
     }
 }
