@@ -1,5 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:image_scanner_example/main.dart';
+
 import 'package:oktoast/oktoast.dart';
 import 'package:photo_manager/photo_manager.dart';
 
@@ -20,7 +21,10 @@ class PhotoProvider extends ChangeNotifier {
 
   bool get needTitle => _needTitle;
 
-  set needTitle(bool needTitle) {
+  set needTitle(bool? needTitle) {
+    if (needTitle == null) {
+      return;
+    }
     _needTitle = needTitle;
     notifyListeners();
   }
@@ -31,6 +35,15 @@ class PhotoProvider extends ChangeNotifier {
 
   set containsEmptyAlbum(bool containsEmptyAlbum) {
     _containsEmptyAlbum = containsEmptyAlbum;
+    notifyListeners();
+  }
+
+  bool _containsPathModified = false;
+
+  bool get containsPathModified => _containsPathModified;
+
+  set containsPathModified(bool containsPathModified) {
+    _containsPathModified = containsPathModified;
     notifyListeners();
   }
 
@@ -56,7 +69,10 @@ class PhotoProvider extends ChangeNotifier {
 
   bool get asc => _asc;
 
-  set asc(bool asc) {
+  set asc(bool? asc) {
+    if (asc == null) {
+      return;
+    }
     _asc = asc;
     notifyListeners();
   }
@@ -80,7 +96,10 @@ class PhotoProvider extends ChangeNotifier {
 
   bool get ignoreSize => _ignoreSize;
 
-  set ignoreSize(bool ignoreSize) {
+  set ignoreSize(bool? ignoreSize) {
+    if (ignoreSize == null) {
+      return;
+    }
     _ignoreSize = ignoreSize;
     notifyListeners();
   }
@@ -103,7 +122,10 @@ class PhotoProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  set notifying(bool notifying) {
+  set notifying(bool? notifying) {
+    if (notifying == null) {
+      return;
+    }
     _notifying = notifying;
     notifyListeners();
   }
@@ -113,19 +135,35 @@ class PhotoProvider extends ChangeNotifier {
     notifyListeners();
   }
 
-  void changeHasAll(bool value) {
+  void changeHasAll(bool? value) {
+    if (value == null) {
+      return;
+    }
     this.hasAll = value;
     notifyListeners();
   }
 
-  void changeOnlyAll(bool value) {
+  void changeOnlyAll(bool? value) {
+    if (value == null) {
+      return;
+    }
     this.onlyAll = value;
     notifyListeners();
   }
 
-  void changeContainsEmptyAlbum(bool value) {
+  void changeContainsEmptyAlbum(bool? value) {
+    if (value == null) {
+      return;
+    }
     this.containsEmptyAlbum = value;
     notifyListeners();
+  }
+
+  void changeContainsPathModified(bool? value) {
+    if (value == null) {
+      return;
+    }
+    this.containsPathModified = value;
   }
 
   void reset() {
@@ -135,11 +173,6 @@ class PhotoProvider extends ChangeNotifier {
 
   Future<void> refreshGalleryList() async {
     final option = makeOption();
-
-    if (option == null) {
-      assert(option != null);
-      return;
-    }
 
     reset();
     var galleryList = await PhotoManager.getAssetPathList(
@@ -159,51 +192,44 @@ class PhotoProvider extends ChangeNotifier {
 
   AssetPathProvider getOrCreatePathProvider(AssetPathEntity pathEntity) {
     pathProviderMap[pathEntity] ??= AssetPathProvider(pathEntity);
-    return pathProviderMap[pathEntity];
+    return pathProviderMap[pathEntity]!;
   }
 
   FilterOptionGroup makeOption() {
-    SizeConstraint sizeConstraint;
-    try {
-      final minW = int.tryParse(minWidth);
-      final maxW = int.tryParse(maxWidth);
-      final minH = int.tryParse(minHeight);
-      final maxH = int.tryParse(maxHeight);
-      sizeConstraint = SizeConstraint(
-        minWidth: minW,
-        maxWidth: maxW,
-        minHeight: minH,
-        maxHeight: maxH,
-        ignoreSize: ignoreSize,
-      );
-    } catch (e) {
-      showToast("Cannot convert your size.");
-      return null;
-    }
-
-    DurationConstraint durationConstraint = DurationConstraint(
-      min: minDuration,
-      max: maxDuration,
-    );
-
     final option = FilterOption(
-      sizeConstraint: sizeConstraint,
-      durationConstraint: durationConstraint,
+      sizeConstraint: SizeConstraint(
+        minWidth: int.tryParse(minWidth) ?? 0,
+        maxWidth: int.tryParse(maxWidth) ?? 100000,
+        minHeight: int.tryParse(minHeight) ?? 0,
+        maxHeight: int.tryParse(maxHeight) ?? 100000,
+        ignoreSize: ignoreSize,
+      ),
+      durationConstraint: DurationConstraint(
+        min: minDuration,
+        max: maxDuration,
+      ),
       needTitle: needTitle,
     );
 
-    final dtCond = DateTimeCond(
+    final createDtCond = DateTimeCond(
       min: startDt,
       max: endDt,
-      asc: asc,
+      ignore: false,
     );
 
     return FilterOptionGroup()
       ..setOption(AssetType.video, option)
       ..setOption(AssetType.image, option)
       ..setOption(AssetType.audio, option)
-      ..dateTimeCond = dtCond
-      ..containsEmptyAlbum = _containsEmptyAlbum;
+      ..createTimeCond = createDtCond
+      ..containsEmptyAlbum = _containsEmptyAlbum
+      ..containsPathModified = _containsPathModified
+      ..addOrderOption(
+        OrderOption(
+          type: OrderOptionType.updateDate,
+          asc: asc,
+        ),
+      );
   }
 
   Future<void> refreshAllGalleryProperties() async {
@@ -242,8 +268,17 @@ class AssetPathProvider extends ChangeNotifier {
     }
   }
 
+  bool refreshing = false;
+
   Future onRefresh() async {
-    await path.refreshPathProperties();
+    if (refreshing) {
+      return;
+    }
+
+    refreshing = true;
+    await path.refreshPathProperties(
+      maxDateTimeToNow: false,
+    );
     final list = await path.getAssetListPaged(0, loadCount);
     page = 0;
     this.list.clear();
@@ -251,9 +286,14 @@ class AssetPathProvider extends ChangeNotifier {
     isInit = true;
     notifyListeners();
     printListLength("onRefresh");
+
+    refreshing = false;
   }
 
   Future<void> onLoadMore() async {
+    if (refreshing) {
+      return;
+    }
     if (showItemCount > path.assetCount) {
       print("already max");
       return;
