@@ -4,6 +4,7 @@
 
 #import <Photos/Photos.h>
 #import "PhotoChangeObserver.h"
+#import "core/PMLogUtils.h"
 
 @interface PhotoChangeObserver () <PHPhotoLibraryChangeObserver>
 @property(nonatomic, strong) FlutterMethodChannel *handler;
@@ -11,8 +12,8 @@
 @end
 
 @implementation PhotoChangeObserver {
-
-
+    
+    
 }
 
 - (void)initWithRegister:(NSObject <FlutterPluginRegistrar> *)registrar {
@@ -30,18 +31,18 @@
         [self.handler invokeMethod:@"change" arguments:@1];
         return;
     }
-
+    
     PHObjectChangeDetails *details = [changeInstance changeDetailsForObject:collection];
     PHObject *object = details.objectAfterChanges;
-    NSLog(@"%@, %@", object.localIdentifier, object.class);
-
+    [PMLogUtils.sharedInstance info: [NSString stringWithFormat:@"%@, %@", object.localIdentifier, object.class]];
+    
     [self.handler invokeMethod:@"change" arguments:@1];
 }
 
 - (PHAssetCollection *)getRecentCollection {
     if (PHPhotoLibrary.authorizationStatus == PHAuthorizationStatusAuthorized) {
         PHFetchResult<PHAssetCollection *> *result = [PHAssetCollection fetchAssetCollectionsWithType:PHAssetCollectionTypeSmartAlbum subtype:PHAssetCollectionSubtypeAny options:nil];
-
+        
         for (PHAssetCollection *collection in result) {
             if (collection.assetCollectionSubtype == PHAssetCollectionSubtypeSmartAlbumUserLibrary) {
                 return collection;
