@@ -7,28 +7,28 @@
 #import <AVFoundation/AVAsset.h>
 
 @implementation PMCacheContainer {
-  NSMutableDictionary<NSString *, PMAssetEntity *> *map;
+    NSMutableDictionary<NSString *, PMAssetEntity *> *map;
 }
 
 - (instancetype)init {
-  self = [super init];
-  if (self) {
-    map = [NSMutableDictionary new];
-  }
-
-  return self;
+    self = [super init];
+    if (self) {
+        map = [NSMutableDictionary new];
+    }
+    
+    return self;
 }
 
 - (void)putAssetEntity:(PMAssetEntity *)entity {
-  map[entity.id] = entity;
+    map[entity.id] = entity;
 }
 
 - (PMAssetEntity *)getAssetEntity:(NSString *)id {
-  return map[id];
+    return map[id];
 }
 
 - (void)clearCache {
-  [map removeAllObjects];
+    [map removeAllObjects];
 }
 
 @end
