@@ -3,6 +3,7 @@
 //
 
 #import "PMFileHelper.h"
+#import "PMImport.h"
 #import <Foundation/Foundation.h>
 #import <Photos/Photos.h>
 
@@ -15,13 +16,19 @@ typedef void (^ChangeIds)(NSArray<NSString *> *);
 @class PMFilterOptionGroup;
 @class PMThumbLoadOption;
 
+#import "PMProgressHandlerProtocol.h"
+#import "PMResultHandler.h"
+#import "PMConvertProtocol.h"
+
 typedef void (^AssetResult)(PMAssetEntity *);
 
 @interface PMManager : NSObject
 
+@property(nonatomic, strong) NSObject <PMConvertProtocol> *converter;
+
 - (BOOL)isAuth;
 
-+ (void)openSetting;
++ (void)openSetting:(NSObject<PMResultHandler>*)result;
 
 - (void)setAuth:(BOOL)auth;
 
@@ -33,11 +40,9 @@ typedef void (^AssetResult)(PMAssetEntity *);
 
 - (void)clearCache;
 
-- (void)getThumbWithId:(NSString *)id1 option:(PMThumbLoadOption *)option resultHandler:(ResultHandler *)handler;
+- (void)getThumbWithId:(NSString *)id1 option:(PMThumbLoadOption *)option resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
 
-- (void)getFullSizeFileWithId:(NSString *)id
-                     isOrigin:(BOOL)isOrigin
-                resultHandler:(ResultHandler *)handler;
+- (void)getFullSizeFileWithId:(NSString *)id isOrigin:(BOOL)isOrigin resultHandler:(NSObject <PMResultHandler> *)handler progressHandler:(NSObject <PMProgressHandlerProtocol> *)progressHandler;
 
 - (PMAssetPathEntity *)fetchPathProperties:(NSString *)id type:(int)type filterOption:(PMFilterOptionGroup *)filterOption;
 
@@ -57,9 +62,11 @@ typedef void (^AssetResult)(PMAssetEntity *);
 
 - (BOOL)existsWithId:(NSString *)assetId;
 
+- (BOOL)entityIsLocallyAvailable:(NSString *)assetId;
+
 - (NSString*)getTitleAsyncWithAssetId: (NSString *) assetId;
 
-- (void)getMediaUrl:(NSString *)assetId resultHandler:(ResultHandler *)handler;
+- (void)getMediaUrl:(NSString *)assetId resultHandler:(NSObject <PMResultHandler> *)handler;
 
 - (NSArray<PMAssetPathEntity *> *)getSubPathWithId:(NSString *)id type:(int)type albumType:(int)albumType option:(PMFilterOptionGroup *)option;
 
@@ -82,4 +89,6 @@ typedef void (^AssetResult)(PMAssetEntity *);
 - (void)requestCacheAssetsThumb:(NSArray *)identifiers option:(PMThumbLoadOption *)option;
 
 - (void)cancelCacheRequests;
+
+- (void)injectModifyToDate:(PMAssetPathEntity *)path;
 @end
