@@ -4,56 +4,145 @@ import '../type.dart';
 
 /// Filter option for get asset.
 ///
-/// 筛选选项, 可以分别设置图片类型和视频类型对应的[FilterOption]
+/// 筛选选项, 可以分别设置图片类型和视频类型对应的 [FilterOption]
 ///
-/// See [FilterOption]
+/// See [FilterOption].
 class FilterOptionGroup {
-  /// An empty option
-  FilterOptionGroup.empty();
+  static final _defaultOrderOption = OrderOption(
+    type: OrderOptionType.updateDate,
+    asc: false,
+  );
 
-  FilterOptionGroup() {
-    setOption(AssetType.image, FilterOption());
-    setOption(AssetType.video, FilterOption());
-    setOption(AssetType.audio, FilterOption());
+  FilterOptionGroup({
+    FilterOption imageOption = const FilterOption(),
+    FilterOption videoOption = const FilterOption(),
+    FilterOption audioOption = const FilterOption(),
+    bool containsEmptyAlbum = false,
+    bool containsPathModified = false,
+    DateTimeCond? createTimeCond,
+    DateTimeCond? updateTimeCond,
+    List<OrderOption> orders = const [],
+  }) {
+    _map[AssetType.image] = imageOption;
+    _map[AssetType.video] = videoOption;
+    _map[AssetType.audio] = audioOption;
+    this.containsEmptyAlbum = containsEmptyAlbum;
+    this.containsPathModified = containsPathModified;
+    this.createTimeCond = createTimeCond ?? DateTimeCond.def();
+    this.updateTimeCond =
+        createTimeCond ?? DateTimeCond.def().copyWith(ignore: true);
+    this.orders.addAll(orders);
   }
+
+  FilterOptionGroup.empty();
 
   final Map<AssetType, FilterOption> _map = {};
 
   /// 是否包含空相册
-  var containsEmptyAlbum = false;
+  ///
+  /// Whether to include an empty album
+  bool containsEmptyAlbum = false;
 
-  DateTimeCond dateTimeCond = DateTimeCond.def();
+  /// If true, the [AssetPathEntity] will return with the last modified time.
+  ///
+  /// See [AssetPathEntity.lastModified].
+  ///
+  /// This is a performance consuming option.
+  /// It's recommended to be true only if you really need it.
+  bool containsPathModified = false;
 
-  FilterOption getOption(AssetType type) {
-    return _map[type];
+  @Deprecated('Please use createTimeCond.')
+  DateTimeCond get dateTimeCond => createTimeCond;
+
+  @Deprecated('Please use createTimeCond.')
+  set dateTimeCond(DateTimeCond dateTimeCond) {
+    createTimeCond = dateTimeCond;
   }
+
+  DateTimeCond createTimeCond = DateTimeCond.def();
+  DateTimeCond updateTimeCond = DateTimeCond.def().copyWith(
+    ignore: true,
+  );
+
+  FilterOption getOption(AssetType type) => _map[type]!;
 
   void setOption(AssetType type, FilterOption option) {
     _map[type] = option;
   }
 
+  final orders = <OrderOption>[];
+
+  void addOrderOption(OrderOption option) {
+    orders.add(option);
+  }
+
   void merge(FilterOptionGroup other) {
-    assert(other != null, 'Cannot merge null.');
     for (final AssetType type in _map.keys) {
-      _map[type] = _map[type]?.merge(other.getOption(type));
+      _map[type] = _map[type]!.merge(other.getOption(type));
     }
     this.containsEmptyAlbum = other.containsEmptyAlbum;
+    this.containsPathModified = other.containsPathModified;
+    this.createTimeCond = other.createTimeCond;
+    this.updateTimeCond = other.updateTimeCond;
+    this.orders
+      ..clear()
+      ..addAll(other.orders);
   }
 
   Map<String, dynamic> toMap() {
     Map<String, dynamic> result = {};
     if (_map.containsKey(AssetType.image)) {
-      result["image"] = _map[AssetType.image].toMap();
+      result['image'] = getOption(AssetType.image).toMap();
     }
     if (_map.containsKey(AssetType.video)) {
-      result["video"] = _map[AssetType.video].toMap();
+      result['video'] = getOption(AssetType.video).toMap();
     }
     if (_map.containsKey(AssetType.audio)) {
-      result["audio"] = _map[AssetType.audio].toMap();
+      result['audio'] = getOption(AssetType.audio).toMap();
+    }
+    result['containsEmptyAlbum'] = containsEmptyAlbum;
+    result['containsPathModified'] = containsPathModified;
+    result['createDate'] = createTimeCond.toMap();
+    result['updateDate'] = updateTimeCond.toMap();
+
+    final ordersList = List<OrderOption>.of(orders);
+    if (ordersList.isEmpty) {
+      ordersList.add(_defaultOrderOption);
     }
 
-    result["date"] = dateTimeCond.toMap();
-    result['containsEmptyAlbum'] = containsEmptyAlbum;
+    result['orders'] = ordersList.map((e) => e.toMap()).toList();
+
+    return result;
+  }
+
+  FilterOptionGroup copyWith({
+    FilterOption? imageOption,
+    FilterOption? videoOption,
+    FilterOption? audioOption,
+    DateTimeCond? createTimeCond,
+    DateTimeCond? updateTimeCond,
+    bool? containsEmptyAlbum,
+    bool? containsPathModified,
+    List<OrderOption>? orders,
+  }) {
+    imageOption ??= _map[AssetType.image];
+    videoOption ??= _map[AssetType.video];
+    audioOption ??= _map[AssetType.audio];
+    containsEmptyAlbum ??= this.containsEmptyAlbum;
+    containsPathModified ??= this.containsPathModified;
+    createTimeCond ??= this.createTimeCond;
+    updateTimeCond ??= this.updateTimeCond;
+    orders ??= this.orders;
+
+    final result = FilterOptionGroup()
+      ..setOption(AssetType.image, imageOption!)
+      ..setOption(AssetType.video, videoOption!)
+      ..setOption(AssetType.audio, audioOption!)
+      ..createTimeCond = createTimeCond
+      ..updateTimeCond = updateTimeCond
+      ..containsEmptyAlbum = containsEmptyAlbum
+      ..containsPathModified = containsPathModified
+      ..orders.addAll(orders);
 
     return result;
   }
@@ -65,10 +154,8 @@ class FilterOptionGroup {
 }
 
 /// Filter option
-///
 /// 筛选选项的详细情况
 class FilterOption {
-  /// See [needTitle], [sizeConstraint] and [durationConstraint]
   const FilterOption({
     this.needTitle = false,
     this.sizeConstraint = const SizeConstraint(),
@@ -86,9 +173,9 @@ class FilterOption {
 
   /// Create a new [FilterOption] with specific properties merging.
   FilterOption copyWith({
-    bool needTitle,
-    SizeConstraint sizeConstraint,
-    DurationConstraint durationConstraint,
+    bool? needTitle,
+    SizeConstraint? sizeConstraint,
+    DurationConstraint? durationConstraint,
   }) {
     return FilterOption(
       needTitle: needTitle ?? this.needTitle,
@@ -99,19 +186,18 @@ class FilterOption {
 
   /// Merge a [FilterOption] into another.
   FilterOption merge(FilterOption other) {
-    assert(other != null, 'Cannot merge null.');
     return FilterOption(
-      needTitle: other.needTitle ?? this.needTitle,
-      sizeConstraint: other.sizeConstraint ?? this.sizeConstraint,
-      durationConstraint: other.durationConstraint ?? this.durationConstraint,
+      needTitle: other.needTitle,
+      sizeConstraint: other.sizeConstraint,
+      durationConstraint: other.durationConstraint,
     );
   }
 
   Map<String, dynamic> toMap() {
     return {
-      "title": needTitle,
-      "size": sizeConstraint.toMap(),
-      "duration": durationConstraint.toMap(),
+      'title': needTitle,
+      'size': sizeConstraint.toMap(),
+      'duration': durationConstraint.toMap(),
     };
   }
 
@@ -139,13 +225,37 @@ class SizeConstraint {
     this.ignoreSize = false,
   });
 
+  SizeConstraint copyWith({
+    int? minWidth,
+    int? maxWidth,
+    int? minHeight,
+    int? maxHeight,
+    bool? ignoreSize,
+  }) {
+    minWidth ??= this.minWidth;
+    maxWidth ??= this.maxHeight;
+
+    minHeight ??= this.minHeight;
+    maxHeight ??= this.maxHeight;
+
+    ignoreSize ??= this.ignoreSize;
+
+    return SizeConstraint(
+      minWidth: minWidth,
+      maxWidth: maxWidth,
+      minHeight: minHeight,
+      maxHeight: maxHeight,
+      ignoreSize: ignoreSize,
+    );
+  }
+
   Map<String, dynamic> toMap() {
     return {
-      "minWidth": minWidth,
-      "maxWidth": maxWidth,
-      "minHeight": minHeight,
-      "maxHeight": maxHeight,
-      "ignoreSize": ignoreSize,
+      'minWidth': minWidth,
+      'maxWidth': maxWidth,
+      'minHeight': minHeight,
+      'maxHeight': maxHeight,
+      'ignoreSize': ignoreSize,
     };
   }
 }
@@ -164,52 +274,79 @@ class DurationConstraint {
 
   Map<String, dynamic> toMap() {
     return {
-      "min": min.inMilliseconds,
-      "max": max.inMilliseconds,
+      'min': min.inMilliseconds,
+      'max': max.inMilliseconds,
     };
   }
 }
 
+/// CreateDate
 class DateTimeCond {
-  static final DateTime zero = DateTime.utc(0);
+  static final DateTime zero = DateTime.fromMillisecondsSinceEpoch(0);
 
   final DateTime min;
   final DateTime max;
-  final bool asc;
+  final bool ignore;
 
   const DateTimeCond({
-    this.min,
-    this.max,
-    this.asc = false, // default desc
-  })  : assert(min != null),
-        assert(max != null),
-        assert(asc != null);
+    required this.min,
+    required this.max,
+    this.ignore = false,
+  });
 
   factory DateTimeCond.def() {
     return DateTimeCond(
       min: zero,
       max: DateTime.now(),
-      asc: false,
     );
   }
 
   DateTimeCond copyWith({
-    final DateTime min,
-    final DateTime max,
-    final bool asc,
+    DateTime? min,
+    DateTime? max,
+    bool? ignore,
   }) {
     return DateTimeCond(
       min: min ?? this.min,
       max: max ?? this.max,
-      asc: asc ?? this.asc,
+      ignore: ignore ?? this.ignore,
     );
   }
 
   Map<String, dynamic> toMap() {
     return {
-      "min": min.millisecondsSinceEpoch,
-      "max": max.millisecondsSinceEpoch,
-      "asc": asc,
+      'min': min.millisecondsSinceEpoch,
+      'max': max.millisecondsSinceEpoch,
+      'ignore': ignore,
     };
   }
 }
+
+class OrderOption {
+  final OrderOptionType type;
+  final bool asc;
+
+  const OrderOption({
+    this.type = OrderOptionType.createDate,
+    this.asc = false,
+  });
+
+  OrderOption copyWith({
+    OrderOptionType? type,
+    bool? asc,
+  }) {
+    return OrderOption(
+      asc: asc ?? this.asc,
+      type: type ?? this.type,
+    );
+  }
+
+  Map<String, dynamic> toMap() {
+    return {
+      'type': type.index,
+      'asc': asc,
+    };
+  }
+}
+
+enum OrderOptionType { createDate, updateDate }
