@@ -7,6 +7,11 @@ class _NotifyManager {
   static const MethodChannel _channel =
       const MethodChannel("top.kikt/photo_manager/notify");
 
+  StreamController<bool> _controller = StreamController.broadcast();
+
+  /// When the notification status change, the listen of stream will be called.
+  Stream<bool> get notifyStream => _controller.stream;
+
   /// callbacks
   var notifyCallback = <ValueChanged<MethodCall>>[];
 
@@ -22,12 +27,14 @@ class _NotifyManager {
   void startHandleNotify() {
     _channel.setMethodCallHandler(_notify);
     _plugin.notifyChange(start: true);
+    _controller.add(true);
   }
 
   /// stop handle notify
   void stopHandleNotify() {
     _plugin.notifyChange(start: false);
     _channel.setMethodCallHandler(null);
+    _controller.add(false);
   }
 
   Future<dynamic> _notify(MethodCall call) async {
@@ -41,6 +48,6 @@ class _NotifyManager {
   }
 
   Future<dynamic> _onChange(MethodCall methodCall) async {
-    notifyCallback.forEach((callback) => callback?.call(methodCall));
+    notifyCallback.toList().forEach((callback) => callback.call(methodCall));
   }
 }
