@@ -4,21 +4,28 @@ class ConvertUtils {
   static List<AssetPathEntity> convertPath(
     Map data, {
     int type = 0,
-    DateTime dt,
-    FilterOptionGroup optionGroup,
+    FilterOptionGroup? optionGroup,
   }) {
     List<AssetPathEntity> result = [];
 
     List list = data["data"];
 
     for (final Map item in list) {
-      final entity = AssetPathEntity(filterOption: optionGroup)
+      final entity = AssetPathEntity()
         ..id = item["id"]
         ..name = item["name"]
         ..typeInt = type
         ..isAll = item["isAll"]
         ..assetCount = item["length"]
-        ..albumType = (item["albumType"] ?? 1);
+        ..albumType = (item["albumType"] ?? 1)
+        ..filterOption = optionGroup ?? FilterOptionGroup();
+
+      final int? modifiedDate = item['modified'];
+
+      if (modifiedDate != null) {
+        entity.lastModified =
+            DateTime.fromMillisecondsSinceEpoch(modifiedDate * 1000);
+      }
 
       result.add(entity);
     }
@@ -31,49 +38,46 @@ class ConvertUtils {
 
     List list = data["data"];
     for (final Map item in list) {
-      final entity = AssetEntity()
-        ..id = item["id"]
-        ..createDtSecond = item["createDt"]
-        ..width = item["width"]
-        ..height = item["height"]
-        ..orientation = (item["orientation"] ?? 0)
-        ..duration = item["duration"]
-        ..modifiedDateSecond = item["modifiedDt"]
-        ..typeInt = item["type"]
-        ..longitude = item["lng"]
-        ..latitude = item["lat"]
-        ..title = item["title"]
-        ..relativePath = item['relativePath'];
-
-      result.add(entity);
+      final asset = _convertMapToAsset(item);
+      if (asset != null) {
+        result.add(asset);
+      }
     }
 
     return result;
   }
 
-  static AssetEntity convertToAsset(Map map) {
-    if (map == null) {
-      return null;
-    }
-
-    Map data = map["data"];
-
+  static AssetEntity? convertToAsset(Map? map) {
+    final Map? data = map?['data'];
     if (data == null) {
       return null;
     }
 
-    final entity = AssetEntity()
-      ..id = data["id"]
-      ..createDtSecond = data["createDt"]
-      ..width = data["width"]
-      ..height = data["height"]
-      ..duration = data["duration"]
-      ..modifiedDateSecond = data["modifiedDt"]
-      ..typeInt = data["type"]
-      ..longitude = data["lng"]
-      ..latitude = data["lat"]
-      ..title = data["title"];
+    return _convertMapToAsset(data);
+  }
 
-    return entity;
+  static AssetEntity? _convertMapToAsset(Map? data) {
+    if (data == null) {
+      return null;
+    }
+
+    final result = AssetEntity(
+      id: data['id'],
+      typeInt: data['type'],
+      width: data['width'],
+      height: data['height'],
+      duration: data['duration'] ?? 0,
+      orientation: data['orientation'] ?? 0,
+      isFavorite: data['favorite'] ?? false,
+      title: data['title'],
+      createDtSecond: data['createDt'],
+      modifiedDateSecond: data['modifiedDt'],
+      relativePath: data['relativePath'],
+      latitude: data['lat'],
+      longitude: data['lng'],
+      mimeType: data['mimeType'],
+    );
+
+    return result;
   }
 }
